begin factoring out common code to be shared between GCS/S3
