+++ pypicloud/storage/object_store.py
@@ -1,0 +1,87 @@
+""" Store packages in S3 """
+import posixpath
+from binascii import hexlify
+
+import logging
+
+from hashlib import md5
+from pyramid.settings import asbool
+
+from .base import IStorage
+from pypicloud.util import get_settings
+
+
+LOG = logging.getLogger(__name__)
+
+
+class ObjectStoreStorage(IStorage):
+
+    """ Storage backend base class containing code that is common between
+        supported object stores (S3 / GCS)
+    """
+    test = False
+
+    def __init__(self, request=None, bucket=None, expire_after=None,
+                 bucket_prefix=None, prepend_hash=None, redirect_urls=None,
+                 sse=None, object_acl=None, storage_class=None, region_name=None,
+                 public_url=False, **kwargs):
+        super(S3Storage, self).__init__(request, **kwargs)
+        self.bucket = bucket
+        self.expire_after = expire_after
+        self.bucket_prefix = bucket_prefix
+        self.prepend_hash = prepend_hash
+        self.redirect_urls = redirect_urls
+        self.sse = sse
+        self.object_acl = object_acl
+        self.storage_class = storage_class
+        self.region_name = region_name
+        self.public_url = public_url
+
+
+    @classmethod
+    def get_bucket(cls, bucket_name, settings):
+        raise NotImplementedError
+
+
+    @classmethod
+    def configure(cls, settings):
+        kwargs = super(S3Storage, cls).configure(settings)
+        kwargs['expire_after'] = int(settings.get('storage.expire_after',
+                                                  60 * 60 * 24))
+        kwargs['bucket_prefix'] = settings.get('storage.prefix', '')
+        kwargs['prepend_hash'] = asbool(settings.get('storage.prepend_hash',
+                                                     True))
+        kwargs['sse'] = sse = settings.get('storage.server_side_encryption')
+        if sse not in [None, 'AES256', 'aws:kms']:
+            LOG.warn("Unrecognized value %r for 'storage.sse'. See "
+                     "https://boto3.readthedocs.io/en/latest/reference/services/s3.html#S3.Object.put "
+                     "for more details", sse)
+        kwargs['object_acl'] = settings.get('storage.object_acl', None)
+        kwargs['storage_class'] = storage_class = settings.get('storage.storage_class')
+        kwargs['redirect_urls'] = asbool(settings.get('storage.redirect_urls',
+                                                      False))
+        bucket_name = settings.get('storage.bucket')
+        if bucket_name is None:
+            raise ValueError("You must specify the 'storage.bucket'")
+        kwargs['bucket'] = cls.get_bucket(bucket_name, settings)
+
+        kwargs['region_name'] = config_settings.get('region_name')
+        kwargs['public_url'] = asbool(settings.get('storage.public_url'))
+        return kwargs
+
+    def calculate_path(self, package):
+        """ Calculates the path of a package """
+        path = package.name + '/' + package.filename
+        if self.prepend_hash:
+            m = md5()
+            m.update(package.filename.encode('utf-8'))
+            prefix = hexlify(m.digest()).decode('utf-8')[:4]
+            path = prefix + '/' + path
+        return path
+
+    def get_path(self, package):
+        """ Get the fully-qualified bucket path for a package """
+        if 'path' not in package.data:
+            filename = self.calculate_path(package)
+            package.data['path'] = self.bucket_prefix + filename
+        return package.data['path']
